requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
jinja2==3.1.6
python-dotenv==1.0.1
httpx==0.28.1
pydantic==2.11.3
